add check userId

+++ src/controllers/comments.controller.ts
@@ -17,6 +17,13 @@ export const getComments = async (
 export const updateComments = async (
     req:RequestWithParamsAndBody<{id: string},CommentInputModel>,
     res:Response) => {
+    const comment = await commentsQueryRepository.findById(req.params.id);
+    if(!comment){
+        return res.sendStatus(Statuses.NOT_FOUND)
+    }
+    if(comment.commentatorInfo.userId !== req.user!.id){
+        return res.sendStatus(Statuses.FORBIDDEN)
+    }
     const isUpdated = await commentsService.update(req.params.id,req.body);
     if(!isUpdated){
         return res.sendStatus(Statuses.NOT_FOUND)
@@ -27,6 +34,13 @@ export const updateComments = async (
 export const deleteComments = async (
     req:RequestWithParams<{id: string}>,
     res:Response) => {
+    const comment = await commentsQueryRepository.findById(req.params.id);
+    if(!comment){
+        return res.sendStatus(Statuses.NOT_FOUND)
+    }
+    if(comment.commentatorInfo.userId !== req.user!.id){
+        return res.sendStatus(Statuses.FORBIDDEN)
+    }
     const isDeleted = await commentsService.delete(req.params.id);
     if(!isDeleted){
         return res.sendStatus(Statuses.NOT_FOUND)
